ConnectionManager: refactor event processing and sychronisation
This change introduces an event queue for serialisation of all
CM-related event handling. In general, state changes, associated
notifications, and other event processing, are performed without
the CM locked, to avoid the deadlocks that have been prevalent in
earlier releases. However, transport state changes that require
a synchronous change in CM behaviour - that is, transport
disconnections - are handled synchronously so that the transport
is not used when it is no longer viable.